Engine: prevent fp-exception during NaN checks
NaN check routine in Image does fp manipulations that trigger
fp-exception due to NaN presence, but it's normal for a routine
that try to detect NaN.
So this patch removes the fp-exception trap during the check.
Note: in current code the trap is enabled only if NDEBUG is not
defined (so mainly in debug version). But as this code is not
aware of that and as the trap can be enabled outside of this code
for any reason, the disabling is forced for any versions that make
sense.